misc/HtmlTestSuite/results/test-1117.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="resources/utils.js"/>

</head><body><input autofocus="" id="i1"/>
<input autofocus="" id="i2"/>
<script>
"use strict";

promise_test(async () =&gt; {
  const input1 = document.querySelector("#i1");
  const input2 = document.querySelector("#i2");
  input1.remove();
  input2.parentNode.insertBefore(input1, input2);

  await waitUntilStableAutofocusState();
  assert_equals(document.activeElement, input2);
}, 'The second autofocus element wins if the first autofocus element was ' +
   'disconnected and reconnected before flushing the autofocus candidates.');
</script>
</body></html>
